Clean up unused _anySetter
