+++ src/main/java/com/fasterxml/jackson/databind/deser/BeanDeserializerFactory.java
@@ -544,11 +544,8 @@ ClassUtil.name(propName)));
 
         // Also, do we have a fallback "any" setter?
         AnnotatedMember anySetter = beanDesc.findAnySetterAccessor();
-        AnnotatedMember creatorPropWithAnySetter = _findCreatorPropWithAnySetter(ctxt, creatorProps);
         if (anySetter != null) {
             builder.setAnySetter(constructAnySetter(ctxt, beanDesc, anySetter));
-        } else if (creatorPropWithAnySetter != null) {
-            builder.setAnySetter(constructAnySetter(ctxt, beanDesc, creatorPropWithAnySetter));
         } else {
             // 23-Jan-2018, tatu: although [databind#1805] would suggest we should block
             //   properties regardless, for now only consider unless there's any setter...
@@ -662,23 +659,6 @@ ClassUtil.name(propName)));
                 builder.addProperty(prop);
             }
         }
-    }
-
-    private AnnotatedMember _findCreatorPropWithAnySetter(DeserializationContext ctxt, SettableBeanProperty[] creatorProps) {
-        if (creatorProps != null) {
-            AnnotationIntrospector ai = ctxt.getAnnotationIntrospector();
-            for (SettableBeanProperty prop : creatorProps) {
-                AnnotatedMember m = prop.getMember();
-                if (m != null) {
-                    Boolean hasAnySetter = ai.hasAnySetter(m);
-                    if (hasAnySetter != null && hasAnySetter) {
-                        return prop.getMember();
-                    }
-                }
-
-            }
-        }
-        return null;
     }
 
     private boolean _isSetterlessType(Class<?> rawType) {
@@ -839,35 +819,6 @@ ClassUtil.name(name), ((AnnotatedParameter) m).getIndex());
                     valueType, null, mutator,
                     PropertyMetadata.STD_OPTIONAL);
 
-        } else if (mutator instanceof AnnotatedParameter){
-            AnnotatedParameter af = (AnnotatedParameter) mutator;
-            // get the type from the content type of the map object
-            JavaType fieldType = af.getType();
-            // 31-Jul-2022, tatu: Not just Maps any more but also JsonNode, so:
-            if (fieldType.isMapLikeType()) {
-                fieldType = resolveMemberAndTypeAnnotations(ctxt, mutator, fieldType);
-                keyType = fieldType.getKeyType();
-                valueType = fieldType.getContentType();
-                prop = new BeanProperty.Std(PropertyName.construct("stuff"),
-                    fieldType, null, mutator, PropertyMetadata.STD_OPTIONAL);
-            } else if (fieldType.hasRawClass(JsonNode.class)
-                || fieldType.hasRawClass(ObjectNode.class)) {
-                fieldType = resolveMemberAndTypeAnnotations(ctxt, mutator, fieldType);
-                // Deserialize is individual values of ObjectNode, not full ObjectNode, so:
-                valueType = ctxt.constructType(JsonNode.class);
-                prop = new BeanProperty.Std(PropertyName.construct(mutator.getName()),
-                    fieldType, null, mutator, PropertyMetadata.STD_OPTIONAL);
-
-                // Unlike with more complicated types, here we do not allow any annotation
-                // overrides etc but instead short-cut handling:
-                return SettableAnyProperty.constructForJsonNodeField(ctxt,
-                    prop, mutator, valueType,
-                    ctxt.findRootValueDeserializer(valueType));
-            } else {
-                return ctxt.reportBadDefinition(beanDesc.getType(), String.format(
-                    "Unsupported type for any-setter: %s -- only support `Map`s, `JsonNode` and `ObjectNode` ",
-                    ClassUtil.getTypeDescription(fieldType)));
-            }
         } else if (isField) {
             AnnotatedField af = (AnnotatedField) mutator;
             // get the type from the content type of the map object
