+++ src/main/java/com/fasterxml/jackson/databind/deser/impl/PropertyBasedCreator.java
@@ -205,8 +205,7 @@ public final class PropertyBasedCreator
 
     public Object build(DeserializationContext ctxt, PropertyValueBuffer buffer) throws IOException
     {
-        Object bean = _valueInstantiator.createFromObjectWith(ctxt,
-                _allProperties, buffer);
+        Object bean = buildSimple(ctxt, buffer);
         // returning null isn't quite legal, but let's let caller deal with that
         if (bean != null) {
             // Object Id to handle?
